+++ aws-java-sdk-eks/src/main/java/com/amazonaws/services/eks/AmazonEKSAsync.java
@@ -26,9 +26,9 @@ import com.amazonaws.services.eks.model.*;
  * </p>
  * <p>
  * <p>
- * Amazon Elastic Container Service for Kubernetes (Amazon EKS) is a managed service that makes it easy for you to run
- * Kubernetes on AWS without needing to stand up or maintain your own Kubernetes control plane. Kubernetes is an
- * open-source system for automating the deployment, scaling, and management of containerized applications.
+ * Amazon Elastic Kubernetes Service (Amazon EKS) is a managed service that makes it easy for you to run Kubernetes on
+ * AWS without needing to stand up or maintain your own Kubernetes control plane. Kubernetes is an open-source system
+ * for automating the deployment, scaling, and management of containerized applications.
  * </p>
  * <p>
  * Amazon EKS runs up-to-date versions of the open-source Kubernetes software, so you can use all the existing plugins